requirements.txt
FlightRadarAPI
kafka-python
flightradar_client
aiohttp
numpy